Named the three basic initial choices
Added initial list of three (3) potential legal framework options, namely: 
> (1) Classic Corporate Model
> (2) BLT Federated Model
> (3) Democratic Model

Evidently Jay has a well thought out option that may be ready to adopt, but there is no current link to this option.  Assuming the additional option is provided in a timely manner it will be included and evaluated as a fourth (4th) option.